reorganize registration client mod
